+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## Release 0.2.0
+
+**Features**
+
+- Included a few missed removed functions
+- Included removed types
+- Included an option to scan for deprecated functions in stdlib9
+
+**Bugfixes**
+
+**Known Issues**
+
 ## Release 0.1.0
 
 **Features**